feat: add BuildMessage component, files restructuring and ui fixes
