+++ components/BuildMessage.tsx
@@ -1,0 +1,16 @@
+import { Fragment } from "react";
+
+function BuildMessage(lines: string[]) {
+  return lines.map((line, index) =>
+    index < lines.length - 1 ? (
+      <Fragment key={index}>
+        {line}
+        <br />
+      </Fragment>
+    ) : (
+      <Fragment key={index}>{line}</Fragment>
+    ),
+  );
+}
+
+export default BuildMessage;
