+++ components/MysteryBox/MysteryBoxHub.tsx
@@ -4,9 +4,9 @@ import MysteryBoxIcon from "@/public/images/validation-mystery-box.png";
 import { EMysteryBoxCategory } from "@/types/mysteryBox";
 import React, { useState } from "react";
 
-import { InfoIcon } from "../Icons/InfoIcon";
-import InfoDrawer from "../InfoDrawer/InfoDrawer";
-import MysteryBoxReward from "../MysteryBoxReward/MysteryBoxReward";
+import { InfoIcon } from "../../app/components/Icons/InfoIcon";
+import InfoDrawer from "../../app/components/InfoDrawer/InfoDrawer";
+import MysteryBoxReward from "../../app/components/MysteryBoxReward/MysteryBoxReward";
 
 interface MysteryBoxHubProps {
   isUserEligibleForValidationReward: boolean;
@@ -25,7 +25,7 @@ function MysteryBoxHub({
         description={`Mystery Boxes are your gateway to exciting rewards! Each box contains surprises like credits or BONK that you can earn through different activities. The more you CHOMP, the more boxes you unlock!`}
       />
       <div className="bg-darkGray50 rounded-2xl p-2">
-        <h1 className="bg-blue-pink-gradient inline-block text-transparent bg-clip-text font-black text-4xl py-2">
+        <h1 className="bg-blue-pink-gradient text-transparent bg-clip-text font-black text-4xl py-2 mx-auo text-center">
           MYSTERY BOXES
         </h1>
         <div className="flex justify-end items-center mb-2 text-gray-400 gap-2 py-2">
