+++ lib/mysteryBox.ts
@@ -1,3 +1,5 @@
+"use server";
+
 import { getJwtPayload } from "@/app/actions/jwt";
 import { SENTRY_FLUSH_WAIT } from "@/app/constants/sentry";
 import { getCurrentUser } from "@/app/queries/user";
